Map fully functional
The map is for user to change location. On default, the map selected the user current location/default location(depending on GPS permission) .User can click on the map to select location. It save the selected location automatically and allow user to see the selected location weather when they reopen the app or disable GPS when they open the app. If the last selected location is not found, the default location is used.
Note: UI not integrated yet.